requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
boto3==1.28.55
boto3-stubs==1.28.55
botocore==1.31.55
botocore-stubs==1.31.57
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
cryptography==40.0.2
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fabric==3.2.2
fastapi==0.103.1
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
invoke==2.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mypy-boto3-dynamodb==1.28.55
mypy-boto3-ec2==1.28.57
mypy-boto3-s3==1.28.55
orjson==3.9.7
paramiko==3.3.1
passlib==1.7.4
prettytable==3.9.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
types-awscrt==0.19.1
types-s3transfer==0.7.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
wrapt==1.15.0
